Mostly Done
CSS is mostly done, and affects most things to let them be in their final state. Gradients affect the sides and top nav bar. The article is now in the middle of the page, above where images will appear. The images only appear when hovering over a link for that view. The entire page now has a brown-ish background, with the paper being all around it.